reorganiza estrutura de components
