--- src/components/app/App.css
@@ -1,28 +1,0 @@
-/**
- * App.css
- *
- * These are the styles for the app component.
- * The component consists of two parts:
- * - Header
- * - Content
- */
-
-.content {
-    width: 80vw;
-    margin: auto;
-}
-
-/**
- * Defines the distance between the header and the content.
- * The distance is applied to screens of at least 600px.
- */
-.header + .content {
-    padding-top: 100px;
-}
-
-/* Extra small devices (phones, 600px and down) */
-@media only screen and (max-width: 600px) {
-    .header + .content {
-        padding-top: calc(180px);
-    }
-}
