--- src/components/hacktoberfest/HacktoberfestPage.js
@@ -1,13 +1,0 @@
-import React from 'react'
-import Ranking from './Ranking'
-
-const HacktoberfestPage = () => {
-  return (
-    <>
-      <h1 style={{ textAlign: 'center' }}>Hacktoberfest 2019</h1>
-      <Ranking />
-    </>
-  )
-}
-
-export default HacktoberfestPage
