+++ src/components/mentorsList/Mentor.css
@@ -1,0 +1,96 @@
+.mentor-grid {
+  display: grid;
+  width: 100%;
+}
+
+/* Extra small devices (phones, 600px and down) */
+@media only screen and (max-width: 600px) {
+  .mentor-grid {
+    grid-template-columns: 1fr;
+  }
+}
+
+/* Small devices (portrait tablets and large phones, 600px and up) */
+@media only screen and (min-width: 600px) {
+  .mentor-grid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+/* Medium devices (landscape tablets, 768px and up) */
+@media only screen and (min-width: 768px) {
+  .mentor-grid {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+/* Large devices (laptops/desktops, 992px and up) */
+@media only screen and (min-width: 992px) {
+  .mentor-grid {
+    grid-template-columns: repeat(4, 1fr);
+  }
+}
+
+/* Extra large devices (large laptops and desktops, 1200px and up) */
+@media only screen and (min-width: 1200px) {
+  .mentor-grid {
+    grid-template-columns: repeat(4, 1fr);
+  }
+}
+
+.img-mentor {
+  border-radius: 100%;
+  width: 100%;
+  filter: grayscale(100%);
+  transition: all 0.7s ease;
+}
+
+.card-mentor:hover .img-mentor {
+  filter: grayscale(0%);
+}
+
+.name-mentor {
+  color: rgb(46, 46, 46);
+  margin-top: 20px;
+  font-size: 18px;
+  font-weight: bolder;
+}
+
+.card-mentor {
+  padding: 30px 50px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  border-radius: 5px;
+}
+
+.card-mentor:hover {
+  background-color: white;
+  box-shadow: 0px 0px 4px rgba(0, 0, 0, 0.11);
+}
+.repository-mentor {
+  background-color: rgb(114, 161, 214);
+  border-radius: 30px;
+  padding: 3px 10px;
+  font-size: 11px;
+  margin: 3px 3px;
+  color: white;
+  font-weight: bolder;
+}
+
+.repository-list-mentor {
+  visibility: hidden;
+  list-style-type: none;
+  display: flex;
+  flex-flow: row wrap;
+  text-align: center;
+  justify-content: center;
+  padding: 0;
+  opacity: 0;
+  transition: opacity 0.7s;
+}
+
+.card-mentor:hover .repository-list-mentor {
+  visibility: visible;
+  opacity: 1;
+}
