--- src/components/projects/OptionBar.css
@@ -1,8 +1,0 @@
-.option-bar {
-    display: flex;
-    justify-content: flex-end;
-}
-
-.sort-select > .label {
-    margin-right: 5px;
-}
