+++ src/components/ranking/Ranking.css
@@ -1,0 +1,41 @@
+.avatar {
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+}
+
+.user {
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.user > * {
+  margin: 0 10px;
+}
+
+.table {
+  width: 100%;
+}
+
+.loading {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 20vh;
+}
+
+th,
+td {
+  border-bottom: 1px solid #ddd;
+  text-align: center;
+}
+
+th {
+  padding: 10px;
+}
+
+.contributor-name:hover {
+  color: #3ab7f1;
+}
