+++ src/components/renderer/MarkDownRenderer.css
@@ -1,0 +1,9 @@
+.markdown a {
+  color: #0366d6;
+}
+
+.markdown code {
+  background-color: rgba(27, 31, 35, 0.05);
+  border-radius: 20%;
+  padding: 0.2em 0.4em;
+}
