+++ src/components/repository/index.js
@@ -1,0 +1,4 @@
+import RepositoryCard from './RepositoryCard'
+import RepositoryGrid from './RepositoryGrid'
+
+export { RepositoryCard, RepositoryGrid }
