--- src/components/team/teamPage.css
@@ -1,29 +1,0 @@
-.team {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-}
-ul {
-    list-style: none;
-}
-
-.member {
-    display: flex;
-    flex-direction: column;
-    margin: 8px;
-    transition: transform 0.2s ease;
-}
-
-.member:hover {
-    transform: scale(1.2);
-}
-
-.member-image {
-    align-self: center;
-    object-fit: cover;
-    max-width: 130px;
-    border-radius: 50%;
-}
-.member-name {
-    align-self: center;
-}
