+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
-import App from './components/app/App'
+import App from './App'
 
 import './index.css'
 import 'blueprint-css/dist/blueprint.min.css'
