+++ src/pages/hacktoberfest/Hacktoberfest.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import Ranking from '../../components/ranking'
+
+const HacktoberfestPage = () => {
+  return (
+    <>
+      <h1 style={{ textAlign: 'center' }}>Hacktoberfest 2019</h1>
+      <Ranking />
+    </>
+  )
+}
+
+export default HacktoberfestPage
