+++ src/pages/home/homePage.css
@@ -1,0 +1,48 @@
+.projects-container {
+  margin-top: 5vw;
+}
+
+.footer {
+  display: flex;
+  justify-content: center;
+  margin-top: 40px;
+}
+
+.loader {
+  display: block;
+  margin: auto;
+  height: 100%;
+  background-image: url('../../../assets/loader.gif');
+  background-size: contain;
+  background-repeat: no-repeat;
+  background-position: center;
+}
+
+.show-more-btn {
+  font-family: 'Lato', monospace;
+  background-color: #21d6a0;
+  border-width: 0;
+  border-radius: 5px;
+  width: 200px;
+  height: 50px;
+  max-width: 100%;
+  margin-bottom: 40px;
+  box-shadow: 0px 1px 5px 0px rgba(0, 0, 0, 0.2),
+    0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 3px 1px -2px rgba(0, 0, 0, 0.12);
+}
+
+.show-more-btn:hover {
+  cursor: pointer;
+  background-color: #1fb488;
+}
+
+.show-more-btn:focus {
+  outline: none;
+}
+
+.show-more-btn-text {
+  color: #ffffff;
+}
+.show-more-btn-text {
+  margin: 0 auto;
+}
