+++ src/pages/index.js
@@ -1,0 +1,6 @@
+import HomePage from './home'
+import HacktoberfestPage from './hacktoberfest/Hacktoberfest'
+import MentorsPage from './mentors/Mentors'
+import JoinUsPage from './community/JoinUs'
+
+export { HomePage, HacktoberfestPage, MentorsPage, JoinUsPage }
