+++ src/pages/mentors/Mentors.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import MentorsList from '../../components/mentorsList'
+
+const MentorsPage = () => <MentorsList />
+
+export default MentorsPage
